+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/case-editor/case-edit-confirm/case-edit-confirm.component.spec.ts
@@ -156,7 +156,6 @@ describe('CaseEditConfirmComponent', () => {
         declarations: [
           CaseEditConfirmComponent,
           MarkdownComponent
-          // Mock
         ],
         providers: [
           {provide: CaseEditComponent, useValue: caseEditCompStub},