+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/case-editor/case-edit/case-edit.component.ts
@@ -4,12 +4,18 @@ import { ActivatedRoute, Params, Router } from '@angular/router';
 import { Observable, Subject } from 'rxjs';
 
 import { ConditionalShowRegistrarService, GreyBarService } from '../../../directives';
-import { CaseEditCaseSubmit, CaseEditGenerateCaseEventData, CaseEditGetNextPage, CaseEditonEventCanBeCompleted,
-  CaseEditSubmitForm, CaseEventData, CaseEventTrigger, CaseField,
-  CaseView, Draft, HttpError, Profile } from '../../../domain';
+import {
+  CaseEditCaseSubmit, CaseEditGenerateCaseEventData, CaseEditGetNextPage,
+  CaseEditSubmitForm,
+  CaseEditonEventCanBeCompleted,
+  CaseEventData, CaseEventTrigger, CaseField,
+  CaseView, Draft, HttpError, Profile
+} from '../../../domain';
 import { Task } from '../../../domain/work-allocation/Task';
-import { FieldsPurger, FieldsUtils, FormErrorService, FormValueService,
-  SessionStorageService, WindowService } from '../../../services';
+import {
+  FieldsPurger, FieldsUtils, FormErrorService, FormValueService,
+  SessionStorageService, WindowService
+} from '../../../services';
 import { Confirmation, Wizard, WizardPage } from '../domain';
 import { EventCompletionParams } from '../domain/event-completion-params.model';
 import { CaseNotifier, WizardFactoryService } from '../services';