+++ projects/ccd-case-ui-toolkit/src/lib/shared/domain/search/search-result-view-item.model.ts
@@ -1,3 +1,4 @@
+// tslint:disable:variable-name
 import { Type } from 'class-transformer';
 import { CaseField } from '../definition';