+++ projects/ccd-case-ui-toolkit/src/lib/shared/pipes/complex/ccd-page-fields.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { CcdPageFieldsPipe } from './ccd-page-fields.pipe';
+
+describe('CddPageFieldsPipe', () => {
+  it('create an instance', () => {
+    const pipe = new CcdPageFieldsPipe();
+    expect(pipe).toBeTruthy();
+  });
+});